refact: refatorado index

+++ admin/index.php
@@ -3,8 +3,10 @@
 require '../init.php';
 
 if (! is_logged_in()) {
-    redirect_to('login.php');
+    redirect_to('../login.php');
 }
+
+$current_user = get_logged_in_user();
 
 $action = isset($_GET['action']) ? $_GET['action'] : '';
 
@@ -16,16 +18,25 @@ switch ($action) {
                     die('Erro ao deletar o Post. Tente novamente.');
                 }
 
-                delete_post($id);
-                redirect_to('admin?action=list-posts');
+                $post = get_post($id);
+                if ($post && can_delete_post($current_user, $post)) {
+                    delete_post($id);
+                    redirect_to('../admin?action=list-posts&success=true');
+                } else {
+                    die('Você não tem permissão para deletar este post.');
+                }
                 die();
             }
 
-            $all_posts = get_all_posts();
+            $all_posts = get_all_posts('', 1, 1000);
             require 'templates/list-posts.php';
             break;
         }
     case 'new-post': {
+            if (!can_create_post($current_user)) {
+                redirect_to('../index.php');
+            }
+
             $error = false;
             $title = '';
             $excerpt = '';
@@ -43,7 +54,7 @@ switch ($action) {
                 } else {
                     insert_post($title, $excerpt, $content);
                     // Redirecionar para a Home
-                    redirect_to('admin?action=list-posts&success=true');
+                    redirect_to('../admin?action=list-posts&success=true');
                 }
             }
